% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_clear_basis}
\alias{hi_solver_clear_basis}
\title{Clear Basis}
\usage{
hi_solver_clear_basis(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
An integer status code (0 = OK).
}
\description{
Invalidates the solver's current basis, allowing presolve to run
on the next call to \code{hi_solver_run}.
}
