% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_get_basis}
\alias{hi_solver_get_basis}
\title{Get Basis}
\usage{
hi_solver_get_basis(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A list with components:
\describe{
  \item{valid}{Logical, whether the basis is valid.}
  \item{col_status}{Integer vector of column basis statuses
    (0=Lower, 1=Basic, 2=Upper, 3=Zero, 4=Nonbasic).}
  \item{row_status}{Integer vector of row basis statuses.}
}
}
\description{
Retrieves the current simplex basis from the solver.
}
