% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_get_ranging}
\alias{hi_solver_get_ranging}
\title{Get Ranging (Sensitivity Analysis)}
\usage{
hi_solver_get_ranging(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A list with components:
\describe{
  \item{valid}{Logical, whether the ranging data is valid.}
  \item{col_cost_up, col_cost_dn}{Lists with value, objective, in_var, ou_var
    vectors for cost ranging (length num_col).}
  \item{col_bound_up, col_bound_dn}{Lists for column bound ranging.}
  \item{row_bound_up, row_bound_dn}{Lists for row bound ranging (length num_row).}
}
}
\description{
Computes ranging information for the solved LP. The solver must have
an optimal LP solution before calling this function.
}
