% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_get_run_time}
\alias{hi_solver_get_run_time}
\title{Get Run Time}
\usage{
hi_solver_get_run_time(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A numeric value (seconds).
}
\description{
Returns the elapsed time of the last solver run.
}
