% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_postsolve}
\alias{hi_solver_postsolve}
\title{Postsolve}
\usage{
hi_solver_postsolve(solver, col_value, col_dual, row_value, row_dual)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{col_value}{Numeric vector of column values.}

\item{col_dual}{Numeric vector of column duals.}

\item{row_value}{Numeric vector of row values.}

\item{row_dual}{Numeric vector of row duals.}
}
\value{
An integer status code (0 = OK).
}
\description{
Performs postsolve using a given solution from the presolved model.
}
