% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_presolve}
\alias{hi_solver_presolve}
\title{Presolve}
\usage{
hi_solver_presolve(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
An integer status code (0 = OK).
}
\description{
Runs presolve on the current model without solving.
}
