% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_read_model}
\alias{hi_solver_read_model}
\title{Read Model from File}
\usage{
hi_solver_read_model(solver, filename)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{filename}{Path to the model file.}
}
\value{
An integer status code (0 = OK).
}
\description{
Reads a model from a file (MPS or LP format).
}
