% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_set_start}
\alias{hi_solver_set_start}
\title{Set Starting Solution for the Solver}
\usage{
hi_solver_set_start(solver, val)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{val}{A numeric vector of values for the variables.
Missing values can be used to specify that particular
variables should have their starting values calculated automatically.}
}
\value{
An integer value indicating the status.
}
\description{
This function sets the starting solution for the solver.
}
\examples{
solver <- example_solver()
hi_solver_set_start(solver, c(0, NA, NA))

}
