% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_version}
\alias{hi_solver_version}
\title{Get HiGHS Version}
\usage{
hi_solver_version(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A character string with the HiGHS version.
}
\description{
Returns the version string of the bundled HiGHS solver.
}
