% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{new_solver}
\alias{new_solver}
\title{Create a new solver instance.}
\usage{
new_solver(model)
}
\arguments{
\item{model}{an object of class \code{"highs_model"}}
}
\value{
A new solver instance.
}
\description{
This function initializes a new Highs solver instance using the provided model pointer.
}
\examples{
model <- example_model()
new_solver(model)
}
