% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_clear_model}
\alias{solver_clear_model}
\title{Clear Model Data}
\usage{
solver_clear_model(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
The solver instance with cleared model data.
}
\description{
This function clears only the model data from the solver instance.
}
\examples{
solver <- example_solver()
solver_clear_model(solver)

}
