% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_clear_solver}
\alias{solver_clear_solver}
\title{Clear Solver State}
\usage{
solver_clear_solver(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
The solver instance with cleared solver state.
}
\description{
This function clears the internal solver state while preserving the model.
}
\examples{
solver <- example_solver()
solver_clear_solver(solver)

}
