% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_get_bool_option}
\alias{solver_get_bool_option}
\title{Get Boolean Option Value}
\usage{
solver_get_bool_option(solver, key)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{key}{A character string specifying the option name.}
}
\value{
A logical value.
}
\description{
This function retrieves the value of a boolean solver option.
}
\examples{
solver <- example_solver()
value <- solver_get_bool_option(solver, "parallel")

}
