% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_get_constraint_bounds}
\alias{solver_get_constraint_bounds}
\title{Get Constraint Bounds}
\usage{
solver_get_constraint_bounds(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A list containing lower and upper bounds for all constraints.
}
\description{
This function retrieves the bounds for all constraints in the model.
}
\examples{
solver <- example_solver()
bounds <- solver_get_constraint_bounds(solver)

}
