% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_get_constraint_matrix}
\alias{solver_get_constraint_matrix}
\title{Get Constraint Matrix}
\usage{
solver_get_constraint_matrix(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A sparse matrix representing the constraints.
}
\description{
This function retrieves the constraint matrix of the optimization model.
}
\examples{
solver <- example_solver()
matrix <- solver_get_constraint_matrix(solver)

}
