% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_get_dbl_option}
\alias{solver_get_dbl_option}
\title{Get Double Option Value}
\usage{
solver_get_dbl_option(solver, key)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{key}{A character string specifying the option name.}
}
\value{
A numeric value.
}
\description{
This function retrieves the value of a double precision solver option.
}
\examples{
solver <- example_solver()
value <- solver_get_dbl_option(solver, "time_limit")

}
