% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_get_lp_costs}
\alias{solver_get_lp_costs}
\title{Get Objective Coefficients}
\usage{
solver_get_lp_costs(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A numeric vector of objective coefficients.
}
\description{
This function retrieves the objective coefficients of the linear program.
}
\examples{
solver <- example_solver()
costs <- solver_get_lp_costs(solver)

}
