% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_get_model}
\alias{solver_get_model}
\title{Get Model from Solver}
\usage{
solver_get_model(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
The current optimization model.
}
\description{
This function retrieves the current optimization model from the solver.
}
\examples{
solver <- example_solver()
model <- solver_get_model(solver)

}
