% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_get_num_row}
\alias{solver_get_num_row}
\title{Get Number of Constraints}
\usage{
solver_get_num_row(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
An integer representing the number of constraints.
}
\description{
This function returns the number of constraints (rows) in the optimization model.
}
\examples{
solver <- example_solver()
n_constraints <- solver_get_num_row(solver)

}
