% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_get_vartype}
\alias{solver_get_vartype}
\title{Get Variable Types}
\usage{
solver_get_vartype(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A character vector of variable types.
}
\description{
This function retrieves the type (continuous, integer, etc.) of all variables.
}
\examples{
solver <- example_solver()
types <- solver_get_vartype(solver)

}
