% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_infinity}
\alias{solver_infinity}
\title{Get Solver Infinity Value}
\usage{
solver_infinity()
}
\value{
A numeric value representing infinity in the solver.
}
\description{
This function returns the value that the solver uses to represent infinity.
}
\examples{
inf <- solver_infinity()

}
