% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_info}
\alias{solver_info}
\title{Get Solver Information}
\usage{
solver_info(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A list containing solver information.
}
\description{
This function retrieves detailed information about the solver's state and performance.
}
\examples{
solver <- example_solver()
info <- solver_info(solver)

}
