% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_set_option}
\alias{solver_set_option}
\title{Set Solver Option}
\usage{
solver_set_option(solver, key, value)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{key}{A character string specifying the option name.}

\item{value}{The value to set for the option.}
}
\value{
The solver instance with the updated option.
}
\description{
This function sets a solver option using a key-value pair.
}
\examples{
solver <- example_solver()
solver_set_option(solver, "time_limit", 100)

}
