% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_solution}
\alias{solver_solution}
\title{Get Solution}
\usage{
solver_solution(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A list containing the solution information.
}
\description{
This function retrieves the solution from the solver after optimization.
}
\examples{
solver <- example_solver()
solver_run(solver)
solution <- solver_solution(solver)

}
