% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_status}
\alias{solver_status}
\title{Get Solver Status}
\usage{
solver_status(solver)
}
\arguments{
\item{solver}{An object of class "highs_solver".}
}
\value{
A status code indicating the solver state.
}
\description{
This function returns the current status of the solver.
}
\examples{
solver <- example_solver()
solver_run(solver)
status <- solver_status(solver)

}
