% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{solver_write_model}
\alias{solver_write_model}
\title{Write Model to File}
\usage{
solver_write_model(solver, filename)
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{filename}{A character string specifying the output file path.}
}
\value{
Invisible NULL.
}
\description{
This function writes the current optimization model to a file.
}
\examples{
solver <- example_solver()
solver_write_model(solver, "model.mps")

}
