% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver-api.R
\name{hi_solver_set_options}
\alias{hi_solver_set_options}
\title{Set Solver Option}
\usage{
hi_solver_set_options(solver, kwargs = list())
}
\arguments{
\item{solver}{An object of class "highs_solver".}

\item{kwargs}{A named list of options to set.}
}
\value{
The solver instance with the updated options.
}
\description{
This function sets a solver option using a key-value pair.
}
\examples{
solver <- example_solver()
hi_solver_set_options(solver, list(time_limit = 100))

}
